server/requirements.txt
flask
flask-restful
flask-httpauth
pyyaml
python-telegram-bot[job-queue]
psutil
gunicorn
